+++ frontend/src/models/Auth.types.ts
@@ -1,0 +1,12 @@
+export type RegisterUserRequestDto = {
+  name: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+};
+
+export type UserResponseDto = {
+  id: string;
+  name: string;
+  email: string;
+};